Refactor some functions into utils module
